Added SumMolecularEnergy, ghost atoms, and arbitrary charge on atoms.
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%402255

+++ mpqc/src/lib/chemistry/qc/wfn/wfn.h
@@ -82,7 +82,7 @@ class Wavefunction: public MolecularEnergy {
                            const RefSCMatrix& orbs,
                            double* orbval = 0);
 
-    int charge();
+    double charge();
     virtual int nelectron() = 0;
 
     // returns the so density
